requirements.txt
# Minimal deps; you can add more later
streamlit>=1.39
langchain>=0.3
chromadb>=0.5
sentence-transformers>=3.0
faiss-cpu>=1.8
pydantic>=2
python-dotenv>=1.0
tqdm>=4.66
